Refactor BaseModule > Module into BaseModule > SafeBaseModule > Module.
