+++ pax/_src/core/module.py
@@ -6,8 +6,9 @@ import jax
 import jax.numpy as jnp
 import jax.tree_util
 
-from .base import BaseModule, EmptyNode
+from .base import EmptyNode
 from .module_and_value import module_and_value
+from .safe_module import SafeBaseModule
 from .threading_local import allow_mutation
 
 T = TypeVar("T", bound="Module")
@@ -16,17 +17,17 @@ TreeDef = Any
 
 
 def parameters_method(trainable_attributes: Iterable[str], submodules=True):
-    """Return a `parameters()` method."""
+    """Return a `parameters` method."""
 
-    def _parameters_method(self: T) -> T:
+    def _parameters(self: T) -> T:
         mod = self.apply_submodules(lambda x: x.parameters())
         for name in mod.pytree_attributes:
             value = getattr(mod, name)
             if submodules:
                 leaves, _ = jax.tree_flatten(
-                    value, is_leaf=lambda x: isinstance(x, BaseModule)
+                    value, is_leaf=lambda x: isinstance(x, Module)
                 )
-                has_submod = any(mod for mod in leaves if isinstance(mod, BaseModule))
+                has_submod = any(mod for mod in leaves if isinstance(mod, Module))
                 if not (has_submod or name in trainable_attributes):
                     mod = mod.replace(**{name: EmptyNode()})
             else:
@@ -34,24 +35,10 @@ def parameters_method(trainable_attributes: Iterable[str], submodules=True):
                     mod = mod.replace(**{name: EmptyNode()})
         return mod
 
-    return _parameters_method
+    return _parameters
 
 
-def update_pytree(mod: T, *, other: T) -> T:
-    """Use non-EmptyNode leaves from other."""
-
-    def _select_fn(leaf_x, leaf_y):
-        if isinstance(leaf_y, EmptyNode):
-            return leaf_x
-        else:
-            return leaf_y
-
-    is_empty = lambda x: isinstance(x, EmptyNode)
-    new_mod = jax.tree_map(_select_fn, mod, other, is_leaf=is_empty)
-    return new_mod
-
-
-class Module(BaseModule):
+class Module(SafeBaseModule):
     """The base class for all PAX modules.
 
     Example:
@@ -66,6 +53,7 @@ class Module(BaseModule):
     """
 
     _name: Optional[str] = None
+    _training: bool
 
     def __init__(self, name: Optional[str] = None):
         """Initialize module.
@@ -76,6 +64,7 @@ class Module(BaseModule):
         """
         super().__init__()
         self._name = name
+        self._training = True
 
     @property
     def name(self):
@@ -93,19 +82,6 @@ class Module(BaseModule):
         False
         """
         return self._training
-
-    def apply_submodules(self: M, func: Callable[..., M]) -> M:
-        """Apply a function to all submodules, recursively."""
-        module = self.copy()
-        submod_fn = lambda x: isinstance(x, Module) and x is not module
-        leaves, treedef = jax.tree_flatten(module, is_leaf=submod_fn)
-        new_leaves = []
-        for value in leaves:
-            if isinstance(value, Module):
-                new_leaves.append(value.apply(func))
-            else:
-                new_leaves.append(value)
-        return jax.tree_unflatten(treedef, new_leaves)
 
     parameters = parameters_method((), submodules=True)
 
@@ -177,70 +153,14 @@ class Module(BaseModule):
         mod.scan_bugs()
         return mod
 
-    def scan_bugs(self: T) -> T:
-        """Scan the module for potential bugs."""
+    def set_attribute(self: T, name, value) -> T:
+        """Create a new module and set attribute."""
+        module = self.copy()
+        with allow_mutation(module):
+            setattr(module, name, value)
+            module.find_and_register_pytree_attributes()
 
-        # scan for shared module/weight.
-        self._assert_not_shared_module()
-        self._assert_not_shared_weight()
-
-        def _scan_field_fn(mod: T) -> T:
-            assert isinstance(mod, Module)
-            # pylint: disable=protected-access
-            mod._scan_fields(mod._class_fields())
-            # pylint: disable=protected-access
-            mod._scan_fields(mod.__dict__.keys())
-            return mod
-
-        self.apply(_scan_field_fn)
-        return self
-
-    def __mod__(self: T, args: Union[Any, Tuple]) -> Tuple[T, Any]:
-        """An alternative to `pax.module_and_value`.
-
-        >>> bn = pax.nn.BatchNorm1D(3)
-        >>> x = jnp.ones((5, 8, 3))
-        >>> bn, y = bn % x
-        >>> bn
-        BatchNorm1D(num_channels=3, ...)
-        """
-        assert callable(self)
-
-        if isinstance(args, tuple):
-            return module_and_value(self)(*args)
-        else:
-            return module_and_value(self)(args)
-
-    def __or__(self: T, other: T) -> T:
-        """Merge two modules.
-
-        >>> a = pax.nn.Linear(2, 2)
-        >>> b = pax.nn.Linear(2, 2)
-        >>> c = a | b
-        >>> d = b | a
-        >>> c == d
-        False
-        """
-        return update_pytree(self, other=other)
-
-    def __invert__(self: T) -> T:
-        return self.parameters()
-
-    def map(self: T, func, *mods) -> T:
-        return jax.tree_map(func, self, *mods)
-
-    def _repr(self, info: Optional[Dict[str, Any]] = None) -> str:
-        name = f"({self.name}) " if self.name is not None else ""
-        cls_name = self.__class__.__qualname__
-        if info is None:
-            return f"{name}{cls_name}"
-        else:
-            lst_info = [f"{k}={v}" for (k, v) in info.items() if v is not None]
-            str_info = ", ".join(lst_info)
-            return f"{name}{cls_name}({str_info})"
-
-    def __repr__(self) -> str:
-        return self._repr()
+        return module
 
     def summary(self, return_list: bool = False) -> Union[str, List[str]]:
         """Summarize a module as a tree of its submodules.
@@ -279,11 +199,146 @@ class Module(BaseModule):
         else:
             return "\n".join(output)
 
-    def set_attribute(self: T, name, value) -> T:
-        """Create a new module and set attribute."""
-        module = self.copy()
-        with allow_mutation(module):
-            setattr(module, name, value)
-            module.find_and_register_pytree_attributes()
+    def map(self: T, func, *mods) -> T:
+        return jax.tree_map(func, self, *mods)
 
-        return module
+    def apply(self: T, apply_fn) -> T:
+        """Apply a function to all submodules.
+
+        >>> def print_param_count(mod):
+        ...     count = sum(jax.tree_leaves(jax.tree_map(jnp.size, mod)))
+        ...     print(f"{count}\t{mod}")
+        ...     return mod
+        ...
+        >>> net = pax.nn.Sequential(pax.nn.Linear(1, 1), jax.nn.relu)
+        >>> net = net.apply(print_param_count)
+        2 Linear(in_dim=1, out_dim=1, with_bias=True)
+        0 Lambda(relu)
+        2 Sequential
+
+        Arguments:
+            apply_fn: a function which inputs a module and outputs a transformed module.
+            check_treedef: check treedef before applying the function.
+        """
+
+        def rec_fn(mod_or_ndarray):
+            if isinstance(mod_or_ndarray, Module):
+                return mod_or_ndarray.apply(apply_fn)
+            else:
+                return mod_or_ndarray
+
+        submodules: List[Module] = self.submodules()
+        new_self = jax.tree_map(
+            rec_fn,
+            self,
+            is_leaf=lambda x: isinstance(x, Module) and (x in submodules),
+        )
+
+        # tree_map already created a copy of self,
+        # hence `apply_fn` is guaranteed to have no side effects.
+        return apply_fn(new_self)
+
+    def apply_submodules(self: M, func: Callable[..., M]) -> M:
+        """Apply a function to all submodules, recursively."""
+        module = self.copy()
+        submod_fn = lambda x: isinstance(x, Module) and x is not module
+        leaves, treedef = jax.tree_flatten(module, is_leaf=submod_fn)
+        new_leaves = []
+        for value in leaves:
+            if isinstance(value, Module):
+                new_leaves.append(value.apply(func))
+            else:
+                new_leaves.append(value)
+        return jax.tree_unflatten(treedef, new_leaves)
+
+    def submodules(self) -> List[T]:
+        """Return a list of submodules."""
+        submod_fn = lambda x: isinstance(x, Module) and x is not self
+        leaves, _ = jax.tree_flatten(self, is_leaf=submod_fn)
+        return [leaf for leaf in leaves if submod_fn(leaf)]
+
+    def scan_bugs(self: T) -> T:
+        """Scan the module for potential bugs."""
+
+        # scan for shared module/weight.
+        self._assert_not_shared_module()
+        self._assert_not_shared_weight()
+
+        def _scan_field_fn(mod: T) -> T:
+            assert isinstance(mod, Module)
+            # pylint: disable=protected-access
+            mod._scan_fields(mod._class_fields())
+            # pylint: disable=protected-access
+            mod._scan_fields(mod.__dict__.keys())
+            return mod
+
+        self.apply(_scan_field_fn)
+        return self
+
+    def __setattr__(self, name: str, value: Any) -> None:
+        self._assert_mutability()
+        super().__setattr__(name, value)
+        self.find_and_register_pytree_attributes()
+
+    def __delattr__(self, name: str) -> None:
+        self._assert_mutability()
+        super().__delattr__(name)
+        self.find_and_register_pytree_attributes()
+
+    def __mod__(self: T, args: Union[Any, Tuple]) -> Tuple[T, Any]:
+        """An alternative to `pax.module_and_value`.
+
+        >>> bn = pax.nn.BatchNorm1D(3)
+        >>> x = jnp.ones((5, 8, 3))
+        >>> bn, y = bn % x
+        >>> bn
+        BatchNorm1D(num_channels=3, ...)
+        """
+        assert callable(self)
+
+        if isinstance(args, tuple):
+            return module_and_value(self)(*args)
+        else:
+            return module_and_value(self)(args)
+
+    def __or__(self: T, other: T) -> T:
+        """Merge two modules.
+
+        >>> a = pax.nn.Linear(2, 2)
+        >>> b = pax.nn.Linear(2, 2)
+        >>> c = a | b
+        >>> d = b | a
+        >>> c == d
+        False
+        """
+        return update_pytree(self, other=other)
+
+    def __invert__(self: T) -> T:
+        return self.parameters()
+
+    def _repr(self, info: Optional[Dict[str, Any]] = None) -> str:
+        name = f"({self.name}) " if self.name is not None else ""
+        cls_name = self.__class__.__qualname__
+        if info is None:
+            return f"{name}{cls_name}"
+        else:
+            lst_info = [f"{k}={v}" for (k, v) in info.items() if v is not None]
+            str_info = ", ".join(lst_info)
+            return f"{name}{cls_name}({str_info})"
+
+    def __repr__(self) -> str:
+        return self._repr()
+
+
+def update_pytree(mod: T, *, other: T) -> T:
+    """Use non-EmptyNode leaves from other."""
+
+    def _select_fn(leaf_x, leaf_y):
+        if isinstance(leaf_y, EmptyNode):
+            return leaf_x
+        else:
+            return leaf_y
+
+    is_empty = lambda x: isinstance(x, EmptyNode)
+    new_mod = jax.tree_map(_select_fn, mod, other, is_leaf=is_empty)
+    return new_mod
